Project/ml-backend/pyproject.toml
[tool.poetry]
name = "ml-backend"
version = "0.1.0"
description = "ML Backend for Memento project"
authors = ["Madiyar Mukushev <madiyar2000@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.74.1"
uvicorn = {extras = ["standard"], version = "^0.17.5"}
python-dotenv = "^0.19.2"
python-multipart = "^0.0.5"
python-magic = "^0.4.25"
python-magic-bin = {version = "^0.4.14", optional = true}

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[tool.poetry.extras]
win_tools = ["python-magic-bin"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
